+++ lib/locator.dart
@@ -8,7 +8,7 @@ import 'package:bb_mobile/features/bitcoin_price/bitcoin_price_locator.dart';
 import 'package:bb_mobile/features/buy/buy_locator.dart';
 import 'package:bb_mobile/features/electrum_settings/electrum_settings_locator.dart';
 import 'package:bb_mobile/features/exchange/exchange_locator.dart';
-import 'package:bb_mobile/features/experimental/broadcast_signed_tx/broadcast_signed_tx_locator.dart';
+import 'package:bb_mobile/features/experimental/broadcast_signed_tx/locator.dart';
 import 'package:bb_mobile/features/experimental/import_watch_only_wallet/import_watch_only_locator.dart';
 import 'package:bb_mobile/features/fund_exchange/fund_exchange_locator.dart';
 import 'package:bb_mobile/features/key_server/key_server_locator.dart';
